requirements.txt
playwright==1.58.0
playwright-stealth==2.0.3
PySide6==6.11.0
qasync==0.28.0
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.10.4
pyyaml==6.0.2
loguru==0.7.3
